fix: Minor fix

+++ data/src/commonMain/kotlin/com/srfmolina/data/db/dao/KrocyItemDao.kt
@@ -17,11 +17,11 @@ interface KrocyItemDao {
     fun getAll(): Flow<List<KrocyItemEntity>>
 
     @Query("SELECT * FROM krocy_items WHERE id = :id LIMIT 1")
-    suspend fun getById(id: Long): KrocyItemEntity?
+    suspend fun getById(id: Int): KrocyItemEntity?
 
     @Delete
     suspend fun delete(item: KrocyItemEntity)
 
     @Query("DELETE FROM krocy_items WHERE id = :id")
-    suspend fun deleteById(id: Long)
+    suspend fun deleteById(id: Int)
 }
